refactor!: Renamed `SpawnConfig` and `RunConfig` to `SpawnOptions` and `RunOptions`.
feat: Users are now able to mount directories using either a existing `Directory` or a `DirectoryInit` which will be used to instantiate a new `Directory`

Release-As: 0.3.0
