+++ package.json
@@ -30,7 +30,7 @@
     "build:dev": "wasm-pack build --dev --target=web --weak-refs --no-pack && rollup -c --environment BUILD:development",
     "dev": "rollup -c -w",
     "test": "web-test-runner --node-resolve --esbuild-target auto --config ./web-dev-server.config.mjs",
-    "doc": "typedoc",
+    "docs": "typedoc",
     "doc:watch": "typedoc --watch",
     "clean": "rimraf dist coverage pkg target"
   },
